+++ manage.py
@@ -6,7 +6,7 @@ import sys
 
 def main():
     os.environ.setdefault('DJANGO_SETTINGS_MODULE',
-                          'project.settings')
+                        'project.settings')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
