+++ photographygal/admin.py
