+++ photographygal/templates/base.html
