+++ project/settings.py
@@ -36,7 +36,7 @@ BASE_DIR = Path(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 SECRET_KEY = env('SECRET_KEY')
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = env("DEBUG")
+DEBUG = True
 
 ALLOWED_HOSTS = []
 
@@ -50,6 +50,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'photographygal',
     
 
     # Third-party
@@ -58,7 +59,7 @@ INSTALLED_APPS = [
 
     # Project-specific
     'users.apps.UsersConfig',
-    'photographygal.apps.PhotographygalConfig',
+    
     
 ]
