+++ project/urls.py
@@ -20,15 +20,16 @@ from photographygal import views
 
 urlpatterns = [
     path('', views.homepage, name='homepage'),
+    #path("photographygal/", include(""photographygal/homepage.html"")),
     path('admin/', admin.site.urls),
     path('accounts/', include('registration.backends.simple.urls')),
 ]
 
 if settings.DEBUG:
     import debug_toolbar
-    urlpatterns = [
-        path('__debug__/', include(debug_toolbar.urls)),
 
+    urlpatterns = [
+        path("__debug__/", include(debug_toolbar.urls)),
         # For django versions before 2.0:
         # url(r'^__debug__/', include(debug_toolbar.urls)),
     ] + urlpatterns
